Non-generic MapperBuilder now sits in its own file
